# docs/量化前端OpenAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stock Project Query API",
    "description": "查询 MongoDB 业务数据和实时行情。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/quant/overview": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Quant Overview",
        "description": "账户资产、盈亏、当日成交汇总和数据状态。省略日期时返回最新已有交易日。",
        "operationId": "quant_overview_api_v1_quant_overview_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "strategy_1",
              "title": "Strategy Id"
            }
          },
          {
            "name": "trade_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quant/strategies/{strategy_id}/overview": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Quant Overview",
        "description": "账户资产、盈亏、当日成交汇总和数据状态。省略日期时返回最新已有交易日。",
        "operationId": "quant_overview_api_v1_quant_strategies__strategy_id__overview_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          },
          {
            "name": "trade_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quant/intraday/latest": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Latest Intraday Result",
        "operationId": "latest_intraday_result_api_v1_quant_intraday_latest_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "strategy_1",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quant/strategies/{strategy_id}/intraday/latest": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Latest Intraday Result",
        "operationId": "latest_intraday_result_api_v1_quant_strategies__strategy_id__intraday_latest_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quant/daily-results/latest": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Latest Daily Result",
        "operationId": "latest_daily_result_api_v1_quant_daily_results_latest_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "strategy_1",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quant/strategies/{strategy_id}/daily-results/latest": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Latest Daily Result",
        "operationId": "latest_daily_result_api_v1_quant_strategies__strategy_id__daily_results_latest_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quant/daily-results/{trade_date}": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Daily Result By Date",
        "description": "匿名名称的完整快照，包括参数、指标、执行过程及逐股账本。",
        "operationId": "daily_result_by_date_api_v1_quant_daily_results__trade_date__get",
        "parameters": [
          {
            "name": "trade_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Trade Date"
            }
          },
          {
            "name": "strategy_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "strategy_1",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quant/strategies/{strategy_id}/daily-results/{trade_date}": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Daily Result By Date",
        "description": "匿名名称的完整快照，包括参数、指标、执行过程及逐股账本。",
        "operationId": "daily_result_by_date_api_v1_quant_strategies__strategy_id__daily_results__trade_date__get",
        "parameters": [
          {
            "name": "trade_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Trade Date"
            }
          },
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/quant/holdings": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "List Quant Holdings",
        "operationId": "list_quant_holdings_api_v1_quant_holdings_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "strategy_1",
              "title": "Strategy Id"
            }
          },
          {
            "name": "trade_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade Date"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{6}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "snapshot_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snapshot Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DailySnapshotResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DailySnapshot"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DailySnapshotResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HoldingPage": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.2",
            "title": "Schema Version",
            "default": "1.2"
          },
          "source_schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Schema Version"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name"
          },
          "trade_date": {
            "type": "string",
            "title": "Trade Date"
          },
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "currency": {
            "type": "string",
            "const": "CNY",
            "title": "Currency",
            "default": "CNY"
          },
          "timezone": {
            "type": "string",
            "const": "Asia/Shanghai",
            "title": "Timezone",
            "default": "Asia/Shanghai"
          },
          "execution_kind": {
            "type": "string",
            "const": "shadow_simulation",
            "title": "Execution Kind",
            "default": "shadow_simulation"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Holding"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "strategy_id",
          "strategy_name",
          "trade_date",
          "snapshot_id",
          "total",
          "page",
          "page_size",
          "items"
        ],
        "title": "HoldingPage"
      },
...
      "OverviewResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Overview"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "OverviewResponse"
      },
...
    }
  }
}